tests/requirements.txt
invoke==0.9.0
coverage==3.7.1
flake8==2.2.5
django-discover-runner==1.0
six==1.8.0

suds==0.4